Separated CognicityServer object from server.js
- Split server specific logic into new CognicityServer object
- Refactored caching from infrastructure requests into CognicityServer
object
