+++ server.js
@@ -2,7 +2,7 @@
 
 //server.js - nodejs server for cognicity framework
 
-//Tomas Holderness January 2014
+// Tomas Holderness January 2014
 
 // Modules
 var path = require('path');
@@ -12,6 +12,7 @@ var cache = require('memory-cache');
 var topojson = require('topojson');
 /** Winston logger module */
 var logger = require('winston');
+var CognicityServer = require('./CognicityServer.js');
 
 // Read in config file from argument or default
 var configFile = ( process.argv[2] ? process.argv[2] : 'config.js' );
@@ -47,6 +48,8 @@ pg.connect(config.pg.conString, function(err, client, done){
 	}
 });
 
+var server = new CognicityServer(config, cache, logger, pg);
+
 // Define a winston stream function we can plug in to express so we can
 // capture its logs along with our own
 var winstonStream = {
@@ -54,6 +57,8 @@ var winstonStream = {
     	logger.info(message.slice(0, -1));
     }
 };
+
+// Setup express logger
 app.use( express.logger( { stream : winstonStream } ) );
 
 // Static file server
@@ -67,15 +72,14 @@ app.all('/'+config.url_prefix+'/data/*', function(req, res, next){
 	next();
 });
 
-
 // Route root path to some place.
 app.get('/', function(req, res){
 	res.redirect('/'+config.root_redirect);
-	});
+});
 
 app.get('/'+config.url_prefix, function(req, res){
 	res.redirect('/'+config.root_redirect);
-	});
+});
 
 //Route empty API path to docs
 app.get('/'+config.url_prefix+'/data/api', function(req, res){
@@ -92,282 +96,6 @@ app.get('/'+config.url_prefix+'/data/', function(req, res){
 	res.redirect('/'+config.url_prefix+'/in/data/');
 });
 
-// Function for database calls
-function dataQuery(pgcon, queryObject, callback){
-	logger.debug("queryObject:"+JSON.stringify(queryObject));
-	pg.connect(pgcon, function(err, client, done){
-		if (err){
-			logger.error("dataQuery: " + JSON.stringify(queryObject) + ", " + err);
-			done();
-			callback({"data":null});
-			return;
-		}
-		
-		client.query(queryObject, function(err, result){
-			if (err){
-				logger.error(JSON.stringify(queryObject) +'\n'+ err);
-				done();
-				callback({"data":null});
-			}
-			else if (result && result.rows){
-				if (result.rows.length === 0){
-					callback({"data":null});
-					done();
-				}
-				else{
-					callback(result.rows);
-					done();
-				}
-			}
-			// something bad happened, return data:null, so client can handle error.
-			else {
-				callback({"data":null});
-				done();
-			}
-		});
-	});
-}
-
-//Cache reports
-function cacheReports(name, data){
-	cache.put(name, data, config.cache_timeout);
-}
-
-function getReports(options, callback){
-
-	// Default parameters for this data
-	// Time parameters hard coded for operation
-	var param = ({
-		start: Math.floor(Date.now()/1000 - 3600), //60 minutes ago.
-		end:  Math.floor(Date.now()/1000), // now
-		limit: config.pg.limit // user adjustable limit
-	});
-
-	for (var key in param){
-		if (options.hasOwnProperty(key)){
-			param[key] = options[key];
-		}
-	}
-
-	// SQL
-	var queryObject = {
-		text: "SELECT 'FeatureCollection' As type, " + 
-				"array_to_json(array_agg(f)) As features " +
-			"FROM (SELECT 'Feature' As type, " + 
-				"ST_AsGeoJSON(lg.the_geom)::json As geometry, " + 
-				"row_to_json( " + 
-					"(SELECT l FROM " +
-						"(SELECT pkey, " +
-						"created_at at time zone 'ICT' created_at, " +
-						"text) " +
-					" As l) " +
-				") As properties " +
-				"FROM " + config.pg.tbl_reports + " As lg " +
-				"WHERE created_at >= to_timestamp($1) AND " +
-					"created_at <= to_timestamp($2) " +
-				"ORDER BY created_at DESC LIMIT $3" + 
-			" ) As f ;",
-		values: [
-            param.start,
-            param.end,
-            param.limit
-		]
-	};
-
-	// Call data query
-	dataQuery(config.pg.conString, queryObject, callback);
-}
-
-// Unconfirmed reports
-function getUnConfirmedReports(options, callback){
-
-	// Default parameters for this data
-	var param = ({
-		start: Math.floor(Date.now()/1000 - 3600), //60 minutes ago.
-		end:  Math.floor(Date.now()/1000), //now
-		limit: config.pg.uc_limit //user adjustable limit
-	});
-
-	for (var key in param){
-		if (options.hasOwnProperty(key)){
-			param[key] = options[key];
-		}
-	}
-
-	// SQL
-	var queryObject = {
-		text: "SELECT 'FeatureCollection' As type, " +
-				"array_to_json(array_agg(f)) As features " +
-			"FROM (SELECT 'Feature' As type, " +
-				"ST_AsGeoJSON(lg.the_geom)::json As geometry, " +
-				"row_to_json( " +
-					"(SELECT l FROM " +
-						"(SELECT pkey) " +
-					"As l) " +
-				") As properties " +
-				"FROM " + config.pg.tbl_reports_unconfirmed + " As lg " +
-				"WHERE created_at >= to_timestamp($1) AND " +
-					"created_at <= to_timestamp($2) " +
-				"ORDER BY created_at DESC LIMIT $3" +
-			" ) As f ;",
-		values: [
-            param.start,
-            param.end,
-            param.limit
-		]
-	};
-
-	// Call data query
-	dataQuery(config.pg.conString, queryObject, callback);
-}
-
-// Function to count unconfirmed reports within given polygon layer (e.g. wards)
-function getCountByArea(options, callback){
-
-	// Default parameters for this data
-	var param = ({
-		start: Math.floor(Date.now()/1000 - 3600), //60 minutes ago
-		end:  Math.floor(Date.now()/1000), // now
-		point_layer_uc: config.pg.tbl_reports_unconfirmed, // unconfirmed reports
-		point_layer: config.pg.tbl_reports, //confirmed reports
-		polygon_layer: config.pg.tbl_polygon_0 // smallest scale polygon table
-	});
-
-	for (var key in param){
-		if (options.hasOwnProperty(key)){
-			param[key] = options[key];
-		}
-	}
-	// SQL
-	// Note that references to tables were left unparameterized as these cannot be passed by user
-	var queryObject = {
-		text: "SELECT 'FeatureCollection' AS type, " +
-				"array_to_json(array_agg(f)) AS features " +
-			"FROM (SELECT 'Feature' AS type, " +
-				"ST_AsGeoJSON(lg.the_geom)::json As geometry," +
-				"row_to_json( " +
-					"(SELECT l FROM " +
-						"(SELECT lg.pkey, " +
-							"lg.area_name as level_name, " +
-							"lg.sum_count as count " +
-						") AS l " +
-					") " +
-				") AS properties " +
-				"FROM ( " +
-					"SELECT c1.pkey, " +
-						"c1.area_name, " +
-						"c1.the_geom, " +
-						"c1.count+c2.count sum_count " +
-					"FROM ( " +
-						"SELECT p1.pkey, " +
-							"p1.area_name, " +
-							"p1.the_geom, " +
-							"COALESCE(count.count,0) count " +
-						"FROM " + param.polygon_layer + " AS p1 " +
-						"LEFT OUTER JOIN ( " +
-							"SELECT b.pkey, " +
-								"count(a.pkey) " +
-							"FROM " + param.point_layer_uc + " a, " + 
-								param.polygon_layer + " b " +
-							"WHERE ST_WITHIN(a.the_geom, b.the_geom) AND " +
-								"a.created_at >=to_timestamp($1) AND " +
-								"a.created_at <= to_timestamp($2) " +
-							"GROUP BY b.pkey " +
-						") as count " +
-						"ON (p1.pkey = count.pkey) " +
-					") as c1, ( " +
-						"SELECT p1.pkey, " +
-							"COALESCE(count.count,0) count  " +
-						"FROM " + param.polygon_layer + " AS p1 " +
-						"LEFT OUTER JOIN( " +
-							"SELECT b.pkey, " +
-								"count(a.pkey) " +
-							"FROM " + param.point_layer + " a, " + 
-								param.polygon_layer + " b " +
-							"WHERE ST_WITHIN(a.the_geom, b.the_geom) AND " +
-								"a.created_at >= to_timestamp($1) AND " +
-								"a.created_at <= to_timestamp($2) " +
-								"GROUP BY b.pkey) as count " +
-						"ON (p1.pkey = count.pkey) " +
-					") as c2 " +
-					"WHERE c1.pkey=c2.pkey " +
-					"ORDER BY pkey " +
-				") AS lg " +
-			") AS f;",
-		values: [
-		    param.start,
-		    param.end
-		]
-	};
-
-	// Call data query
-	dataQuery(config.pg.conString, queryObject, callback);
-}
-
-// Function to cache aggregates of report counts per polygon area
-function cacheCount(name, data){
-	cache.put(name, data, config.cache_timeout);
-}
-
-//Sum of confirmed and unconfirmed aggregates from archive
-function getHistoricalCountByArea(end_time, callback){
-
-	var queryObject = {
-		text: "SELECT 'FeatureCollection' AS type, " +
-				"array_to_json(array_agg(f)) AS features " +
-			"FROM (SELECT 'Feature' AS type, " +
-				"ST_AsGeoJSON(lg.the_geom)::json AS geometry, " +
-				"row_to_json( " +
-					"(SELECT l FROM " +
-						"(SELECT lg.level_name, lg.sum_count, lg.start_time, lg.end_time) AS l " +
-					") " +
-				") AS properties FROM (" +
-					"SELECT a.area_name as level_name, " +
-						"a.the_geom, " +
-						"b.count+c.count sum_count, " +
-						"b.start_time, " +
-						"b.end_time " +
-					"FROM jkt_rw_boundary a, " +
-						"rw_count_reports_confirmed b, " +
-						"rw_count_reports_unconfirmed c " +
-					"WHERE b.rw_pkey = a.pkey AND " +
-						"b.rw_pkey = c.rw_pkey AND " +
-						"b.end_time = to_timestamp($1) AND " +
-						"c.end_time = to_timestamp($1) " +
-				") AS lg " +
-			") AS f;",
-		values: [
-		    end_time
-		]
-	};
-
-	//Call data query
-	dataQuery(config.pg.conString, queryObject, callback);
-}
-
-function getInfrastructure(name, callback){
-
-	var queryObject = {
-		text: "SELECT 'FeatureCollection' AS type, " +
-				"array_to_json(array_agg(f)) AS features " +
-			"FROM (SELECT 'Feature' AS type, " +
-				"ST_AsGeoJSON(lg.the_geom)::json AS geometry, " +
-				"row_to_json( " +
-					"(SELECT l FROM (SELECT name) as l) " +
-				") AS properties " +
-				"FROM " + config.pg.infrastructure_tbls[name] + " AS lg " +
-			") AS f;",
-		values: []		
-	};
-
-	//Call data query
-	dataQuery(config.pg.conString, queryObject, callback);
-}
-
-//Function to cache infrastructure on first call, no timeout set
-function cacheInfrastructure(name, data){
-	cache.put(name, data);
-}
 
 if (config.data === true){
 
@@ -378,8 +106,8 @@ if (config.data === true){
 		var opts = {};
 
 		if (cache.get('reports') === null){
-			getReports(opts, function(data){
-				cacheReports('reports', data);
+			server.getReports(opts, function(data){
+				server.cacheReports('reports', data);
 				writeGeoJSON(res, data[0], req.param('format'));
 			});
 		}
@@ -395,8 +123,8 @@ if (config.data === true){
 			var opts = {};
 
 			if (cache.get('reports_unconfirmed') === null){
-				getUnConfirmedReports(opts, function(data){
-					cacheReports('reports_unconfirmed', data);
+				server.getUnConfirmedReports(opts, function(data){
+					server.cacheReports('reports_unconfirmed', data);
 					writeGeoJSON(res, data[0], req.param('format'));
 				});
 			}
@@ -441,8 +169,8 @@ if (config.data === true){
 					}
 					// Get data from db and update cache.
 					if (cache.get('count_'+level+'_'+hours) === null){
-						getCountByArea({polygon_layer:tbl,start:start}, function(data){
-							cacheCount('count_'+level+'_'+hours, data);
+						server.getCountByArea({polygon_layer:tbl,start:start}, function(data){
+							server.cacheCount('count_'+level+'_'+hours, data);
 
 							// Write data
 							writeGeoJSON(res, data[0], req.param('format'));
@@ -464,7 +192,7 @@ if (config.data === true){
 			else {
 				end_time = 'NULL';
 			}
-			getHistoricalCountByArea(end_time, function(data){
+			server.getHistoricalCountByArea(end_time, function(data){
 				writeGeoJSON(res, data[0], req.param('format'));
 			});
 		});
@@ -472,45 +200,32 @@ if (config.data === true){
 
 	//Data route for waterways infrastructure
 	app.get('/'+config.url_prefix+'/data/api/v1/infrastructure/waterways', function(req, res){
-		if (cache.get('waterways') === null){
-			getInfrastructure('waterways', function(data){
-				cacheInfrastructure('waterways', data);
-				writeGeoJSON(res, data[0], req.param('format'));
-			});
-		}
-		else {
-			writeGeoJSON(res, cache.get('waterways')[0], req.param('format'));
-		}
+		server.getInfrastructure('waterways', function(data){
+			writeGeoJSON(res, data[0], req.param('format'));
+		});
 	});
 
 	//Data route for pump stations
 	app.get('/'+config.url_prefix+'/data/api/v1/infrastructure/pumps', function(req, res){
-		if (cache.get('pumps') === null){
-			getInfrastructure('pumps', function(data){
-				cacheInfrastructure('pumps', data);
-				writeGeoJSON(res, data[0], req.param('format'));
-			});
-		}
-		else {
-			writeGeoJSON(res, cache.get('pumps')[0], req.param('format'));
-		}
+		server.getInfrastructure('pumps', function(data){
+			writeGeoJSON(res, data[0], req.param('format'));
+		});
 	});
 
 	//Data route for floodgates
 	app.get('/'+config.url_prefix+'/data/api/v1/infrastructure/floodgates', function(req, res){
-		if (cache.get('floodgates') === null){
-			getInfrastructure('floodgates', function(data){
-				cacheInfrastructure('floodgates', data);
-				writeGeoJSON(res, data[0], req.param('format'));
-			});
-		}
-		else {
-			writeGeoJSON(res, cache.get('floodgates')[0], req.param('format'));
-		}
+		server.getInfrastructure('floodgates', function(data){
+			writeGeoJSON(res, data[0], req.param('format'));
+		});
 	});
 }
 
-// Function to return GeoJson or TopoJson data to stream
+// 404 handling
+app.use(function(req, res, next){
+  res.send('Error 404 - Page not found', 404);
+});
+
+//Function to return GeoJson or TopoJson data to stream
 function writeGeoJSON(res, data, format){
 	if (format === 'topojson' && data.features !== null){
 		//Clone the object because topojson edits in place.
@@ -532,11 +247,6 @@ function writeGeoJSON(res, data, format){
 		}
 	}
 }
-
-// 404 handling
-app.use(function(req, res, next){
-  res.send('Error 404 - Page not found', 404);
-});
 
 // Use the PORT environment variable (e.g. from AWS Elastic Beanstalk) or use 8081 as the default port
 logger.info( "Application starting, listening on port " + config.port );
